+++ mithril-client/src/type_alias.rs
@@ -37,6 +37,12 @@ pub use mithril_common::messages::CertificateListItemMessageMetadata as MithrilC
 pub use mithril_common::messages::SignerWithStakeMessagePart as MithrilSigner;
 
 cfg_unstable! {
+    /// A Cardano stake distribution.
+    pub use mithril_common::messages::CardanoStakeDistributionMessage as CardanoStakeDistribution;
+
+    /// List item of Cardano stake distributions.
+    pub use mithril_common::messages::CardanoStakeDistributionListItemMessage as CardanoStakeDistributionListItem;
+
     pub use mithril_common::messages::CardanoTransactionsProofsMessage as CardanoTransactionsProofs;
 
     pub use mithril_common::messages::CardanoTransactionsSetProofMessagePart as CardanoTransactionsSetProof;
@@ -46,11 +52,9 @@ cfg_unstable! {
     pub use mithril_common::messages::VerifyCardanoTransactionsProofsError;
 
     /// A snapshot that allow to know up to which [point of time][common::CardanoDbBeacon] Mithril have certified Cardano transactions.
-    ///
     pub use mithril_common::messages::CardanoTransactionSnapshotMessage as CardanoTransactionSnapshot;
 
     /// List item of a Cardano transaction snapshot.
-    ///
     pub use mithril_common::messages::CardanoTransactionSnapshotListItemMessage as CardanoTransactionSnapshotListItem;
 }
 
@@ -62,5 +66,6 @@ pub mod common {
     };
     cfg_unstable! {
         pub use mithril_common::entities::{ChainPoint, TransactionHash, SlotNumber, BlockHash, BlockNumber};
+        pub use mithril_common::entities::{StakeDistribution};
     }
 }